\name{rsnpsingene.cor}
\docType{data}
\alias{rsnpsingene.cor}
\title{SNP correlation matrix RData}
\description{
  This rsnpsingene.cor RData contains SNP correlation matrix
  for each SKATgene computed based on simulated data of 2671 
  exome chip SNPs on chromosome 21.
}
\usage{rsnpsingene.cor}
\format{lists of SNP correlation matrix of 2671 simulated exome chip SNPs on chromosome 21}
\keyword{datasets}