\name{snpinfo}
\docType{data}
\alias{snpinfo}
\title{SNP information RData}
\description{
  This snpinfo RData contains "Name" (SNP name), "Chr" 
  (chromosome number), and "SKATgene" of 2671 exome chip SNPs
     on chromosome 21.
}
\usage{snpinfo}
\format{A matrix containing Name, Chr and SKATgene of 2671 simulated SNPs.}
\keyword{datasets}